+++ app/src/main/java/hr/dtakac/prognoza/apimodel/PlaceData.kt
@@ -6,9 +6,9 @@ data class PlaceResponse(
     @SerializedName("place_id")
     val id: String,
     @SerializedName("lat")
-    val latitude: Float,
+    val latitude: Double,
     @SerializedName("lon")
-    val longitude: Float,
+    val longitude: Double,
     @SerializedName("display_name")
     val displayName: String
 )
